Move handlers to controllers, add header / footer
Now user packages like the users and videos packages as containers for
helper functions for those models, and move the logic of the handlers
into controllers. This helps keep the two types of logic from polluting
each other, and also helps keep module separation cleaner.

Also added logic for appending a header and footer to rendered
templates. Now instead of rendering each template directly, they each
define a "body" template which is then included from the root template.
This is done transparently to the call to render, however, and render
still acts as if it is directly rendering the named template.